Sphinx extension: properly parse first line of code blocks
Sphinx adds a blank span tag to the beginning of each code block
now, which broke the feature that let you click on a line to
execute it. Now the extension considers the content of the first
two elements in each line of code when deciding whether a line is
a Python statement or output, avoiding this problem.